Ajuste en el tema abound
- Configuración del tema.
- Organización de directorios del tema.
