+++ protected/views/example_customer/_form.php
