+++ protected/views/example_customer/_search.php
