+++ protected/views/example_customer/_view.php
