+++ protected/views/example_customer/admin.php
