+++ protected/views/example_customer/create.php
