+++ protected/views/example_customer/index.php
