+++ protected/views/example_customer/view.php
