+++ protected/views/example_personalInfo/_form.php
