+++ protected/views/example_personalInfo/_search.php
