+++ protected/views/example_personalInfo/_view.php
