+++ protected/views/example_personalInfo/admin.php
