+++ protected/views/example_personalInfo/index.php
