+++ protected/views/example_personalInfo/view.php
