--- protected/views/layouts/column1.php
@@ -1,6 +1,0 @@
-<?php /* @var $this Controller */ ?>
-<?php $this->beginContent('//layouts/main'); ?>
-<div id="content">
-	<?php echo $content; ?>
-</div><!-- content -->
-<?php $this->endContent(); ?>
