+++ protected/views/site/pages/about.php
@@ -1,12 +1,133 @@
 <?php
 /* @var $this SiteController */
 
-$this->pageTitle=Yii::app()->name . ' - About';
+$this->pageTitle=Yii::app()->name . ' - Grid';
 $this->breadcrumbs=array(
-	'About',
+	'Grid',
 );
 ?>
-<h1>About</h1>
+<h1>Grid</h1>
 
-<p>This is a "static" page. You may change the content of this page
-by updating the file <code><?php echo __FILE__; ?></code>.</p>
+<div class="row-fluid">
+	<div class="span12 well">
+	  <h3>Heading</h3>
+	  <p>Donec id elit non mi porta gravida at eget metus. Fusce dapibus, tellus ac cursus commodo, tortor mauris condimentum nibh, ut fermentum massa justo sit amet risus. Etiam porta sem malesuada magna mollis euismod. Donec sed odio dui. </p>
+	  <p><a class="btn" href="#">View details &raquo;</a></p>
+	</div><!--/span-->
+</div><!--/row-->
+
+<div class="row-fluid">
+	<div class="span6 well">
+	  <h3>Heading</h3>
+	  <p>Donec id elit non mi porta gravida at eget metus. Fusce dapibus, tellus ac cursus commodo, tortor mauris condimentum nibh, ut fermentum massa justo sit amet risus. Etiam porta sem malesuada magna mollis euismod. Donec sed odio dui. </p>
+	  <p><a class="btn" href="#">View details &raquo;</a></p>
+	</div><!--/span-->
+	<div class="span6 well">
+	  <h3>Heading</h3>
+	  <p>Donec id elit non mi porta gravida at eget metus. Fusce dapibus, tellus ac cursus commodo, tortor mauris condimentum nibh, ut fermentum massa justo sit amet risus. Etiam porta sem malesuada magna mollis euismod. Donec sed odio dui. </p>
+	  <p><a class="btn" href="#">View details &raquo;</a></p>
+	</div><!--/span-->
+</div><!--/row-->
+
+<div class="row-fluid">
+	<div class="span4 well">
+	  <h3>Heading</h3>
+	  <p>Donec id elit non mi porta gravida at eget metus. Fusce dapibus, tellus ac cursus commodo, tortor mauris condimentum nibh, ut fermentum massa justo sit amet risus. Etiam porta sem malesuada magna mollis euismod. Donec sed odio dui. </p>
+	  <p><a class="btn" href="#">View details &raquo;</a></p>
+	</div><!--/span-->
+	<div class="span4 well">
+	  <h3>Heading</h3>
+	  <p>Donec id elit non mi porta gravida at eget metus. Fusce dapibus, tellus ac cursus commodo, tortor mauris condimentum nibh, ut fermentum massa justo sit amet risus. Etiam porta sem malesuada magna mollis euismod. Donec sed odio dui. </p>
+	  <p><a class="btn" href="#">View details &raquo;</a></p>
+	</div><!--/span-->
+	<div class="span4 well">
+	  <h3>Heading</h3>
+	  <p>Donec id elit non mi porta gravida at eget metus. Fusce dapibus, tellus ac cursus commodo, tortor mauris condimentum nibh, ut fermentum massa justo sit amet risus. Etiam porta sem malesuada magna mollis euismod. Donec sed odio dui. </p>
+	  <p><a class="btn" href="#">View details &raquo;</a></p>
+	</div><!--/span-->
+</div><!--/row-->
+
+<div class="row-fluid">
+	<div class="span3 well">
+	  <h3>Heading</h3>
+	  <p>Donec id elit non mi porta gravida at eget metus. Fusce dapibus, tellus ac cursus commodo, tortor mauris condimentum nibh, ut fermentum massa justo sit amet risus. Etiam porta sem malesuada magna mollis euismod. Donec sed odio dui. </p>
+	  <p><a class="btn" href="#">View details &raquo;</a></p>
+	</div><!--/span-->
+	<div class="span3 well">
+	  <h3>Heading</h3>
+	  <p>Donec id elit non mi porta gravida at eget metus. Fusce dapibus, tellus ac cursus commodo, tortor mauris condimentum nibh, ut fermentum massa justo sit amet risus. Etiam porta sem malesuada magna mollis euismod. Donec sed odio dui. </p>
+	  <p><a class="btn" href="#">View details &raquo;</a></p>
+	</div><!--/span-->
+	<div class="span3 well">
+	  <h3>Heading</h3>
+	  <p>Donec id elit non mi porta gravida at eget metus. Fusce dapibus, tellus ac cursus commodo, tortor mauris condimentum nibh, ut fermentum massa justo sit amet risus. Etiam porta sem malesuada magna mollis euismod. Donec sed odio dui. </p>
+	  <p><a class="btn" href="#">View details &raquo;</a></p>
+	</div><!--/span-->
+	<div class="span3 well">
+	  <h3>Heading</h3>
+	  <p>Donec id elit non mi porta gravida at eget metus. Fusce dapibus, tellus ac cursus commodo, tortor mauris condimentum nibh, ut fermentum massa justo sit amet risus. Etiam porta sem malesuada magna mollis euismod. Donec sed odio dui. </p>
+	  <p><a class="btn" href="#">View details &raquo;</a></p>
+	</div><!--/span-->
+</div><!--/row-->
+
+<div class="row-fluid">
+	<div class="span8 well">
+	  <h3>Heading</h3>
+	  <p>Donec id elit non mi porta gravida at eget metus. Fusce dapibus, tellus ac cursus commodo, tortor mauris condimentum nibh, ut fermentum massa justo sit amet risus. Etiam porta sem malesuada magna mollis euismod. Donec sed odio dui. </p>
+	  <p><a class="btn" href="#">View details &raquo;</a></p>
+	</div><!--/span-->
+	<div class="span4 well">
+	  <h3>Heading</h3>
+	  <p>Donec id elit non mi porta gravida at eget metus. Fusce dapibus, tellus ac cursus commodo, tortor mauris condimentum nibh, ut fermentum massa justo sit amet risus. Etiam porta sem malesuada magna mollis euismod. Donec sed odio dui. </p>
+	  <p><a class="btn" href="#">View details &raquo;</a></p>
+	</div><!--/span-->
+</div><!--/row-->
+
+<div class="row-fluid">
+	<div class="span4 well">
+	  <h3>Heading</h3>
+	  <p>Donec id elit non mi porta gravida at eget metus. Fusce dapibus, tellus ac cursus commodo, tortor mauris condimentum nibh, ut fermentum massa justo sit amet risus. Etiam porta sem malesuada magna mollis euismod. Donec sed odio dui. </p>
+	  <p><a class="btn" href="#">View details &raquo;</a></p>
+	</div><!--/span-->
+	<div class="span4 well">
+	  <h3>Heading</h3>
+	  <p>Donec id elit non mi porta gravida at eget metus. Fusce dapibus, tellus ac cursus commodo, tortor mauris condimentum nibh, ut fermentum massa justo sit amet risus. Etiam porta sem malesuada magna mollis euismod. Donec sed odio dui. </p>
+	  <p><a class="btn" href="#">View details &raquo;</a></p>
+	</div><!--/span-->
+	<div class="span4 well">
+	  <h3>Heading</h3>
+	  <p>Donec id elit non mi porta gravida at eget metus. Fusce dapibus, tellus ac cursus commodo, tortor mauris condimentum nibh, ut fermentum massa justo sit amet risus. Etiam porta sem malesuada magna mollis euismod. Donec sed odio dui. </p>
+	  <p><a class="btn" href="#">View details &raquo;</a></p>
+	</div><!--/span-->
+</div><!--/row-->
+
+<div class="row-fluid">
+	<div class="span10 well">
+	  <h3>Heading</h3>
+	  <p>Donec id elit non mi porta gravida at eget metus. Fusce dapibus, tellus ac cursus commodo, tortor mauris condimentum nibh, ut fermentum massa justo sit amet risus. Etiam porta sem malesuada magna mollis euismod. Donec sed odio dui. </p>
+	  <p><a class="btn" href="#">View details &raquo;</a></p>
+	</div><!--/span-->
+	<div class="span2 well">
+	  <h3>Heading</h3>
+	  <p>Donec id elit non mi porta gravida at eget. </p>
+	  <p><a class="btn" href="#">View details &raquo;</a></p>
+	</div><!--/span-->
+</div><!--/row-->
+
+<div class="row-fluid">
+	<div class="span3 well">
+	  <h3>Heading</h3>
+	  <p>Donec id elit non mi porta gravida at eget metus. Fusce dapibus, tellus ac cursus commodo, tortor mauris condimentum nibh, ut fermentum massa justo sit amet risus. Etiam porta sem malesuada magna mollis euismod. Donec sed odio dui. </p>
+	  <p><a class="btn" href="#">View details &raquo;</a></p>
+	</div><!--/span-->
+	<div class="span4 well">
+	  <h3>Heading</h3>
+	  <p>Donec id elit non mi porta gravida at eget metus. Fusce dapibus, tellus ac cursus commodo, tortor mauris condimentum nibh, ut fermentum massa justo sit amet risus. Etiam porta sem malesuada magna mollis euismod. Donec sed odio dui. </p>
+	  <p><a class="btn" href="#">View details &raquo;</a></p>
+	</div><!--/span-->
+	<div class="span5 well">
+	  <h3>Heading</h3>
+	  <p>Donec id elit non mi porta gravida at eget metus. Fusce dapibus, tellus ac cursus commodo, tortor mauris condimentum nibh, ut fermentum massa justo sit amet risus. Etiam porta sem malesuada magna mollis euismod. Donec sed odio dui. </p>
+	  <p><a class="btn" href="#">View details &raquo;</a></p>
+	</div><!--/span-->
+</div><!--/row-->
