+++ protected/views/site/pages/graphs.php
