+++ themes/abound/views/layouts/main.php
@@ -1,14 +1,11 @@
-
 <!DOCTYPE html>
 <html lang="en">
   <head>
     <meta charset="utf-8">
     <title>abound - Yii theme</title>
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <meta name="description" content="Free yii themes, free web application theme">
     <meta name="author" content="Webapplicationthemes.com">
-	<link href='http://fonts.googleapis.com/css?family=Carrois+Gothic' rel='stylesheet' type='text/css'>
-
+    <link href='http://fonts.googleapis.com/css?family=Carrois+Gothic' rel='stylesheet' type='text/css'>
     <!-- Le HTML5 shim, for IE6-8 support of HTML5 elements -->
     <!--[if lt IE 9]>
       <script src="http://html5shim.googlecode.com/svn/trunk/html5.js"></script>
@@ -48,23 +45,17 @@
 	  $cs->registerScriptFile($baseUrl.'/js/styleswitcher.js');
 	?>
   </head>
-
-<body>
-
-<section id="navigation-main">   
-<!-- Require the navigation -->
-<?php require_once('tpl_navigation.php')?>
-</section><!-- /#navigation-main -->
+  <body>
+    <section id="navigation-main">   
+    <?php require_once('tpl_navigation.php') ?>
+    </section>
     
-<section class="main-body">
-    <div class="container-fluid">
-            <!-- Include content pages -->
-            <?php echo $content; ?>
-    </div>
-</section>
-
-<!-- Require the footer -->
-<?php require_once('tpl_footer.php')?>
-
+    <section class="main-body">
+      <div class="container-fluid">
+        <?php echo $content; ?>
+      </div>
+    </section>
+  
+    <?php require_once('tpl_footer.php') ?>
   </body>
 </html>
