+++ themes/abound/views/layouts/tpl_footer.php
@@ -2,7 +2,9 @@
         <div class="subnav navbar navbar-fixed-bottom">
             <div class="navbar-inner">
                 <div class="container">
-                    Designed by <a href="http://www.webapplicationthemes.com" target="_new">webapplicationthemes.com</a>. All Rights Reserved.<br /><small>Powered by <a href="http://www.yiiframework.com" title="Yii Framework" target="_new">Yii Framework</a> and <a href="http://twitter.github.com/bootstrap/" title="Twitter Bootstrap" target="_new">Twitter Bootstrap</a></small>
+                    Elaborado <a href="" target="_new"></a>
+                    <br/>
+                    <small>Powered by <a href="http://www.yiiframework.com" title="Yii Framework" target="_new">Yii Framework</a> and <a href="http://twitter.github.com/bootstrap/" title="Twitter Bootstrap" target="_new">Twitter Bootstrap</a></small>
                 </div>
             </div>
         </div>
