--- themes/abound/views/site/login.php
@@ -1,71 +1,0 @@
-<?php
-/* @var $this SiteController */
-/* @var $model LoginForm */
-/* @var $form CActiveForm  */
-
-$this->pageTitle=Yii::app()->name . ' - Login';
-$this->breadcrumbs=array(
-	'Login',
-);
-?>
-<div class="page-header">
-	<h1>Login <small>to your account</small></h1>
-</div>
-<div class="row-fluid">
-	
-    <div class="span6 offset3">
-<?php
-	$this->beginWidget('zii.widgets.CPortlet', array(
-		'title'=>"Private access",
-	));
-	
-?>
-
-
-
-    <p>Please fill out the following form with your login credentials:</p>
-    
-    <div class="form">
-    <?php $form=$this->beginWidget('CActiveForm', array(
-        'id'=>'login-form',
-        'enableClientValidation'=>true,
-        'clientOptions'=>array(
-            'validateOnSubmit'=>true,
-        ),
-    )); ?>
-    
-        <p class="note">Fields with <span class="required">*</span> are required.</p>
-    
-        <div class="row">
-            <?php echo $form->labelEx($model,'username'); ?>
-            <?php echo $form->textField($model,'username'); ?>
-            <?php echo $form->error($model,'username'); ?>
-        </div>
-    
-        <div class="row">
-            <?php echo $form->labelEx($model,'password'); ?>
-            <?php echo $form->passwordField($model,'password'); ?>
-            <?php echo $form->error($model,'password'); ?>
-            <p class="hint">
-                Hint: You may login with <kbd>demo</kbd>/<kbd>demo</kbd> or <kbd>admin</kbd>/<kbd>admin</kbd>.
-            </p>
-        </div>
-    
-        <div class="row rememberMe">
-            <?php echo $form->checkBox($model,'rememberMe'); ?>
-            <?php echo $form->label($model,'rememberMe'); ?>
-            <?php echo $form->error($model,'rememberMe'); ?>
-        </div>
-    
-        <div class="row buttons">
-            <?php echo CHtml::submitButton('Login',array('class'=>'btn btn btn-primary')); ?>
-        </div>
-    
-    <?php $this->endWidget(); ?>
-    </div><!-- form -->
-
-<?php $this->endWidget();?>
-
-    </div>
-
-</div>
