--- themes/classic/views/.htaccess
@@ -1,1 +1,0 @@
-deny from all
